Move feedback caching functionality implementation to utilities
