+++ packages/frontend/src/pages/FullTestPage.tsx
@@ -2,7 +2,7 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { Layout } from '../Layout';
 import { ConfirmFullTestStart } from '../components/ConfirmFullTestStart';
 import useWebSocket, { ReadyState } from 'react-use-websocket';
-import { useSocketUrl } from '../utilities';
+import { setCachedFeedback, useSocketUrl } from '../utilities';
 import { useState } from 'react';
 import { ListeningQuestionsPage } from './ListeningQuestionsPage';
 import ReadingQuestions from './ReadingQuestionsPage';
@@ -10,7 +10,6 @@ import { Spinner } from '../components/Spinner';
 import { WritingPage } from './WritingPage';
 import { IntermediatePage } from '../components/IntermediatePage';
 import { ToastContainer, toast } from 'react-toastify';
-import { setCachedFeedback } from '../components/AllFeedbacks';
 
 export const FullTestPage = () => {
   let out;
