+++ packages/frontend/src/utilities.ts
@@ -24,3 +24,28 @@ export const useSocketUrl = (): string | undefined => {
   // console.log(`URL is ${import.meta.env.VITE_WEBSOCKET_URL}?idToken=${token}`);
   return `${import.meta.env.VITE_WEBSOCKET_URL}?idToken=${token}`;
 };
+
+////// Full Test Feedback Cache //////
+const getFeedbackKey = (testId: string) => `FeedbackItem-${testId}`;
+
+/**
+ * Pull cached feedback from browser storage
+ *
+ * @param testId  The `id` of the test.
+ * @returns the full test item if it exists, null otherwise
+ */
+export const getCachedFeedback = (testId: string) => {
+  const localData = sessionStorage.getItem(getFeedbackKey(testId));
+  if (localData) console.log('Using cached', { localData });
+  return localData ? JSON.parse(localData) : localData;
+};
+
+/**
+ * Cach feedback in browser storage.
+ *
+ * @param feedback  The feedback to store in cache
+ * @param testId  The `id` of the test.
+ */
+export const setCachedFeedback = (feedback: any, testId: string) => {
+  sessionStorage.setItem(getFeedbackKey(testId), JSON.stringify(feedback));
+};
